+++ base/code/src/main/java/io/almostrealism/expression/Logarithm.java
@@ -24,8 +24,19 @@ import io.almostrealism.lang.LanguageOperations;
 import java.util.List;
 import java.util.OptionalLong;
 
+/**
+ * A natural logarithm expression that computes {@code log(input)}.
+ *
+ * <p>Generates code of the form {@code log(input)}. The factory method cancels
+ * {@code log(exp(x))} back to {@code x}. The derivative is {@code 1 / input}.</p>
+ */
 public class Logarithm extends Expression<Double> {
 
+	/**
+	 * Constructs a natural logarithm expression for the given operand.
+	 *
+	 * @param input the operand whose logarithm is computed
+	 */
 	protected Logarithm(Expression<Double> input) {
 		super(Double.class, input);
 	}
@@ -66,11 +77,22 @@ public class Logarithm extends Expression<Double> {
 		return quotient(target.getShape(), delta, u);
 	}
 
+	/**
+	 * Creates a natural logarithm expression, cancelling {@code log(exp(x))} to {@code x}.
+	 *
+	 * <p>Non-FP operands are widened with {@link Expression#toDouble()} so the emitted
+	 * {@code log()} call always has a floating-point argument — Metal's {@code log()}
+	 * overloads are FP-only and an integer operand triggers a compile ambiguity.</p>
+	 *
+	 * @param input the operand
+	 * @param <T>   the result type (always {@link Double})
+	 * @return the simplified expression or a new {@link Logarithm}
+	 */
 	public static <T> Expression<T> of(Expression input) {
 		if (input instanceof Exp) {
 			return (Expression<T>) input.getChildren().get(0);
 		}
 
-		return (Expression<T>) new Logarithm(input);
+		return (Expression<T>) new Logarithm(input.toDouble());
 	}
 }
